requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
#Django==3.1.3
#django-allauth==0.50.0
#django-celery-results==2.3.1
#django-cors-headers==3.11.0
#django-filter==21.1
#django-rest-auth==0.9.5
#django-rest-knox==4.2.0
#django-seed==0.3.1
#djangorestframework==3.13.1
dlib==19.23.0
Faker==13.12.0
idna==3.3
imagezmq==1.1.1
importlib-metadata==4.11.3
jsonparser==1.0
kombu==5.2.4
Markdown==3.3.6
npm==0.1.1
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
optional-django==0.1.0
packaging==21.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
pyzmq==23.0.0
redis==4.3.1
requests==2.27.1
requests-oauthlib==1.3.1
scipy==1.8.1
six==1.16.0
sqlparse==0.4.2
toposort==1.7
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
zmq==0.0.0
